+++ openmls/src/ciphersuite/aead.rs
@@ -7,7 +7,11 @@ pub(crate) const NONCE_BYTES: usize = 12;
 
 /// AEAD keys holding the plain key value and the AEAD algorithm type.
 #[derive(Serialize, Deserialize)]
-#[cfg_attr(any(feature = "test-utils", test), derive(Clone, PartialEq, Eq))]
+#[cfg_attr(any(feature = "test-utils", test), derive(PartialEq, Eq))]
+#[cfg_attr(
+    any(feature = "test-utils", feature = "virtual-clients-draft", test),
+    derive(Clone)
+)]
 #[cfg_attr(feature = "crypto-debug", derive(Debug))]
 pub struct AeadKey {
     aead_mode: AeadType,