+++ CHANGELOG.md
@@ -10,11 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - [#1972](https://github.com/openmls/openmls/pull/1972): Add APIs for time-based deletion of past epoch secrets, and for setting the past epoch deletion policy for an `MlsGroup`.
+- [#2010](https://github.com/openmls/openmls/pull/2010): Added `MlsGroup::propose_self_update_with_new_signer`, a variant of `propose_self_update` that stages an `Update` proposal carrying a new signature key.
 
 ### Changed
 
 - [#1980](https://github.com/openmls/openmls/pull/1980): Enrich limetime related errors returned during leaf node validation with more information
 - [#1972](https://github.com/openmls/openmls/pull/1972): The method `MlsGroup::max_past_epochs()` now returns an `Option<usize>`, rather than a `usize`
+- [#1963](https://github.com/openmls/openmls/pull/1963): `MessageEncryptionError` is now part of the public API, but only surfaces if the `virtual-clients-draft` feature flag is enabled.
 
 ## 0.8.1 (2026-02-13)